Revert "Recoup simulation on Sky Mote: Fix cluster id feature"
This reverts commit f761a5ba0cde5d60410c40926c16b963681381af.

+++ core/net/ipv6/uip-ds6-nbr.h
@@ -78,7 +78,7 @@ typedef struct uip_ds6_nbr {
   struct uip_packetqueue_handle packethandle;
 #define UIP_DS6_NBR_PACKET_LIFETIME CLOCK_SECOND * 4
 #endif                          /*UIP_CONF_QUEUE_PKT */
-  uint8_t cluster_id;
+  int cluster_id;
 } uip_ds6_nbr_t;
 
 void uip_ds6_neighbors_init(void);
@@ -98,7 +98,7 @@ void uip_ds6_neighbor_periodic(void);
 int uip_ds6_nbr_num(void);
 
 const uip_lladdr_t *uip_ds6_nbr_lladdr_from_ipaddr_for_cluster_id(const uip_ipaddr_t *ipaddr);
-uint8_t uip_ds6_nbr_lookup_for_cluster_id(const uip_ipaddr_t *ipaddr);
+int uip_ds6_nbr_lookup_for_cluster_id(const uip_ipaddr_t *ipaddr);
 
 
 /**
